+++ src/resources/canvas/mod.rs
@@ -3,5 +3,7 @@
 */
 pub mod defs;
 mod brush;
+mod pen;
 
 pub use self::brush::{Brush, BrushT};
+pub use self::pen::{PenT, Pen};